requirements.txt
requests
feedparser
redis
python-dotenv
pillow
